+++ Assets/OxGKit/LoggingSystem/Scripts/Runtime/Logging/Enum/LogLevel.cs
@@ -6,7 +6,7 @@ namespace OxGKit.LoggingSystem
     public enum LogLevel
     {
         Off = 0,
-        Log = 1 << 0,
+        LogDebug = 1 << 0,
         LogInfo = 1 << 1,
         LogWarning = 1 << 2,
         LogError = 1 << 3,